# backend/openapi.yaml
openapi: 3.0.3
info:
  title: GreenLight API
  description: >
    Red comunitaria offline-first para reportar quemas y focos de humo.
    Los reportes pasan por: activo &rarr; confirmado_comunidad &rarr; verificado.
    Documentado con Swagger en /api/docs. Autenticación JWT (Bearer) firmada
    con SECRET_KEY; errores: 401 (token ausente/inválido/expirado), 403
    (sin permiso), 404 (recurso inexistente/invisible), 422 (validación),
    409 (conflicto).
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Desarrollo local
  - url: https://greenlight-api-a487.onrender.com
    description: Producción (Render)
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/registro:
    post:
      tags: [Auth]
      summary: Crea una cuenta de usuario
      operationId: registrar
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroBody'
      responses:
        '201':
          description: Usuario creado (devuelve token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: El correo ya está registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validación fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Inicia sesión y devuelve un access token JWT
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [correo, contraseña]
              properties:
                correo:
                  type: string
                  example: maria@greenlight.test
                contraseña:
                  type: string
                  format: password
                  example: Vecino123!
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales incorrectas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Perfil del usuario autenticado
      operationId: perfil
      responses:
        '200':
          description: Datos del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          description: Token ausente, inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/usuarios:
    get:
      tags: [Auth]
      summary: Lista usuarios (solo coordinador)
      operationId: listarUsuarios
      responses:
        '200':
          description: Listado de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
        '403':
          description: No eres coordinador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Auth]
      summary: Crea un usuario (solo coordinador)
      operationId: crearUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroBody'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
  /api/auth/usuarios/{usuario_id}:
    put:
      tags: [Auth]
      summary: Actualiza nombre/rol/comunidad de un usuario (solo coordinador)
      operationId: actualizarUsuario
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                rol:
                  type: string
                  enum: [usuario, coordinador]
                comunidad_id:
                  type: string
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Auth]
      summary: Elimina un usuario (solo coordinador)
      operationId: eliminarUsuario
      parameters:
        - $ref: '#/components/parameters/UsuarioId'
      responses:
        '200':
          description: Usuario eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UsuarioId:
      name: usuario_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Descripción del error
    Usuario:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        correo:
          type: string
        rol:
          type: string
          enum: [usuario, coordinador]
        comunidad_id:
          type: string
          nullable: true
        comunidad:
          type: string
          nullable: true
        fecha_creacion:
          type: string
          format: date-time
          nullable: true
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        usuario:
          $ref: '#/components/schemas/Usuario'
    RegistroBody:
      type: object
      required: [nombre, correo, contraseña]
      properties:
        nombre:
          type: string
          minLength: 2
        correo:
          type: string
          format: email
        contraseña:
          type: string
          format: password
          minLength: 6
        rol:
          type: string
          enum: [usuario, coordinador]
          default: usuario
        comunidad_id:
          type: string
          nullable: true
...